+++ docs/changelog.md
@@ -8,6 +8,10 @@ nav_order: 2
 
 # Unreleased changes
 
+### Protocol
+
+- Add A_SystemNetworkParameter_Read and A_SystemNetworkParameter_Response APCI service parsing.
+
 ### DPT
 
 - Add DPT 2 definitions